Refactor repository classes
